fix: corrige bloqueio indevido ao sacar todo o saldo do caixa
O saldo do caixa era calculado subtraindo floats (float(entradas) -
float(saidas)), o que podia gerar um valor com erro de arredondamento
ligeiramente menor que o exibido na tela (ex: 66018.81999999999 em vez
de 66018.82). Ao tentar retirar exatamente o saldo mostrado, a
comparação "valor > calcular_saldo()" acusava saldo insuficiente e
bloqueava a retirada.

Agora o saldo e o valor lançado são mantidos em Decimal até o momento
de serializar a resposta, eliminando esse erro de precisão.

+++ app/blueprints/home/routes.py
@@ -1,4 +1,5 @@
 from datetime import datetime, date
+from decimal import Decimal
 
 from flask import Blueprint, render_template
 from sqlalchemy import func
@@ -13,7 +14,7 @@ home_bp = Blueprint("home", __name__)
 def calcular_saldo_caixa():
     entradas = db.session.query(func.coalesce(func.sum(MovimentacaoCaixa.valor), 0)).filter_by(tipo="entrada").scalar()
     saidas = db.session.query(func.coalesce(func.sum(MovimentacaoCaixa.valor), 0)).filter_by(tipo="saida").scalar()
-    return float(entradas) - float(saidas)
+    return float(Decimal(entradas) - Decimal(saidas))
 
 
 @home_bp.route("/home")
